+++ src/helix/endpoints/channels/mod.rs
@@ -37,6 +37,8 @@ pub mod get_followed_channels;
 pub mod get_vips;
 pub mod modify_channel_information;
 pub mod remove_channel_vip;
+#[cfg(feature = "beta")]
+pub mod snooze_next_ad;
 pub mod start_commercial;
 
 #[doc(inline)]
@@ -61,5 +63,7 @@ pub use modify_channel_information::{
 };
 #[doc(inline)]
 pub use remove_channel_vip::{RemoveChannelVipRequest, RemoveChannelVipResponse};
+#[cfg(feature = "beta")]
+pub use snooze_next_ad::{SnoozeNextAdRequest, SnoozedAdSchedule};
 #[doc(inline)]
 pub use start_commercial::{StartCommercial, StartCommercialBody, StartCommercialRequest};